fix: simplify hexdump loop, solving column misalignment problem
Previously, the hexdump loop would always read into a `BUFSIZ`-sized buffer,
and then print lines one-by-one from this buffer.  This caused problems when a
column count is specified (-w parameter) but not a multiple of BUFSIZ.  To
solve this issue, and also simplify the code, instead read only the bytes for
one row at a time.